Fixed float error for 1.20.1

+++ src/main/java/tigeax/customwings/nms/NMSSupport.java
@@ -16,7 +16,7 @@ public class NMSSupport {
             Method handle = entity.getMethod("getHandle");
             Object nmsEntity = handle.invoke(livingEntity);
             Field bodyRotation = nmsEntity.getClass().getField(getBodyRotationField(version));
-            return ((Integer) bodyRotation.get(nmsEntity)).floatValue();
+            return (Float) bodyRotation.get(nmsEntity);
         } catch (Exception e) {
             e.printStackTrace();
             return null;
